requirements.txt
flask
python-dotenv
tiktoken
faiss-cpu
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters
langchain-google-genai